Added sequencer-triggered envelopes (mono only)
All three sequencer step modes are now functional. Auto is the typical functionality, where the sequencer steps forward at a constant pace. Lock-To-Gate (or just Gate) causes the sequencer to step every time a note is pressed, and works in both monophonic and polyphonic mode. Trigger does the opposite: the evenlopes get re-triggered every time the sequencer moves from one step to the next (only works in mono for the moment, as it can lead to orphaned notes in polyphonic mode).